+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-addons/compare/0.3.0...HEAD
 
+### Added
+
+- The control `chaosaddons.controls.repeat` to run an activity multiple times
+
 ## [0.3.0][]
 
 [0.3.0]: https://github.com/chaostoolkit/chaostoolkit-addons/compare/0.2.0...0.3.0